This PR makes compatibility changes for the Windows client.

Merge pull request #24 from leil-io/windows-client-dependency-code

+++ tests/test_suites/ShortSystemTests/test_cs_failure_recovery_speed.sh
@@ -10,9 +10,15 @@ mkdir "${info[mount0]}/dir"
 saunafs setgoal 4 "${info[mount0]}/dir"
 head -c 1M /dev/urandom | tee "${info[mount0]}"/dir/file{0,1,2,3,4} >/dev/null
 
+wait_if_windows
+
 # Stop one of chunkservers which has one copy of each chunk and
 # immediately start overwriting all the chunks.
-kill -s SIGSTOP "$(saunafs_chunkserver_daemon 0 test 2>&1 | sed 's/.*pid: //')"
+if is_windows_system; then
+	saunafs_chunkserver_daemon 0 stop
+else
+	kill -s SIGSTOP "$(saunafs_chunkserver_daemon 0 test 2>&1 | sed 's/.*pid: //')"
+fi
 for i in {0..4}; do
 	( assert_success file-overwrite "${info[mount$i]}/dir/file$i" && touch "$TEMP_DIR/finish$i" & )
 done
@@ -20,4 +26,7 @@ done
 # Expect these tasks to finish not later than after 4 seconds. The timeout is calculated as:
 # 2 s (ma<->cs timeout) + 0.5 s (cl<->cs timeout) + 1 s (client can wait ~1 s between retries)
 assert_success wait_for '(( $(ls "$TEMP_DIR"/finish? 2>/dev/null | wc -l) == 5 ))' '4 seconds'
+
+wait_if_windows
+
 assert_success file-validate "${info[mount0]}/dir/"file*
